Undid some more unnecessary changes
